// order-frontend/src/app/shared/models/order-status.enum.ts
export enum OrderStatus {
  PENDING = 'PENDING',
  PAYMENT_PROCESSING = 'PAYMENT_PROCESSING',
  PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  INVENTORY_PROCESSING = 'INVENTORY_PROCESSING',
  INVENTORY_SUCCESS = 'INVENTORY_SUCCESS',
  INVENTORY_FAILED = 'INVENTORY_FAILED',
  SHIPPING_PROCESSING = 'SHIPPING_PROCESSING',
  SHIPPING_SUCCESS = 'SHIPPING_SUCCESS',
  SHIPPING_FAILED = 'SHIPPING_FAILED',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED'
}